data_extractor/pyproject.toml
[project]
name = "data_extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyodbc",
    "python-dotenv",
    "langchain[google-genai]",
    "pydantic",
    "langchain-chroma",
    "streamlit",
    "nest-asyncio"
]
